feat: jest conversion
