+++ package.json
@@ -46,7 +46,7 @@
     "@rollup/plugin-babel": "^5.2.1",
     "@rollup/plugin-replace": "^2.3.4",
     "@vue/compiler-sfc": "^3.0.2",
-    "@vue/test-utils": "^1.0.0-beta.12",
+    "@vue/test-utils": "^1.1.1",
     "babel-eslint": "^7.1.1",
     "babel-jest": "^26.6.3",
     "chalk": "2.4.2",
@@ -70,6 +70,7 @@
     "rollup-plugin-typescript2": "^0.29.0",
     "rollup-plugin-vue": "^6.0.0-beta.11",
     "shelljs": "0.8.3",
+    "ts-jest": "^26.4.4",
     "typescript": "^4.1.2",
     "vue": "^2.6.12",
     "vue-jest": "4.0.0-rc.0",
