+++ test/app/package.json
@@ -7,14 +7,15 @@
     "build": "vue-cli-service build"
   },
   "dependencies": {
+    "@okta/okta-vue": "*",
     "vue": "^2.6.11",
-    "vue-router": "^3.2.0",
-    "@okta/okta-vue": "*"
+    "vue-router": "^3.2.0"
   },
   "devDependencies": {
     "@vue/cli-plugin-router": "~4.5.0",
     "@vue/cli-plugin-typescript": "~4.5.0",
     "@vue/cli-service": "~4.5.0",
+    "source-map-loader": "^1.1.3",
     "vue-template-compiler": "^2.6.11"
   }
 }
